PythonService/requirements.txt
altgraph==0.17.2
blinker==1.6.2
click==8.1.3
Flask==2.3.2
future==0.18.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.15.2
MarkupSafe==2.1.3
Pillow==9.5.0
PyMuPDF==1.22.3
PyPDF2==3.0.1
reportlab==4.0.4
setuptools==58.0.4
six==1.15.0
typing_extensions==4.6.3
Werkzeug==2.3.4
wheel==0.37.0
zipp==3.15.0